+++ lib/hashie/array.rb
@@ -1,0 +1,11 @@
+module Hashie
+  class Array < ::Array
+    include Hashie::Extensions::Array::PrettyInspect
+    include Hashie::Extensions::RubyVersionCheck
+    with_minimum_ruby('2.3.0') do
+      def dig(*indexes)
+        super(*indexes.map { |idx| Integer(idx) })
+      end
+    end
+  end
+end